Update troubleshooting.rst
Seems to me like those who wrote this think too much like programmers and too little like writers. A good thing I would say, but not in this case.
There is ZERO documentation of log level options in the --log argument when using the ``appimage-builder  --generate`` command. Then only reference to there being an option other than the default ``INFO`` option, is appimage-inspector troubleshooting documentation. This does not, however, help a user that is attempting to debug ``appimage-builder --generate``. After 3 days of trying to debug this, it took me a cold hard guess to pass ``DEBUG`` as a ``--log`` option. Why is this not documented? This option is needed when working with the strace tool or gdb tool as it lets us know at what point appimage-builder is coming to an unexpected halt. 
It took me days to figure this out. Please consider adding documentation in regards to the log options that can be passed. If there is a specific reason why this documentation is excluded, please state that reason, so end users know wtf is going on.